refactor: improve job setup and audit packaging
