+++ apps/api/src/jobs/manager.ts
@@ -1,0 +1,8 @@
+import { jobManager } from '../lib/jobManager';
+
+export class JobManager {
+  private static inst = jobManager;
+  static instance() {
+    return JobManager.inst;
+  }
+}
