+++ apps/api/src/routes/report.consistency.ts
@@ -37,6 +37,7 @@ export const reportConsistencyRoutes: FastifyPluginAsync = async (app) => {
         })
         .safeParse(req.body);
       if (!body.success) return reply.code(400).send({ error: 'Invalid body' });
+      // upsert is optional; invoke only if provided
       provider.upsert?.(body.data.accountId, body.data.date, body.data.net);
       return { ok: true };
     });
