+++ apps/api/src/tests/setup.ts
@@ -1,25 +1,15 @@
-import { afterEach, beforeAll, beforeEach, vi } from 'vitest';
 import { resetBusForTests } from '../lib/bus';
-import { stopAllJobs, getJobManagerForTests, resetJobsForTests } from '../jobs/jobManagerTestHooks';
+import { JobManager } from '../jobs/manager';
+import { resetSchedulerForTests } from '../jobs/scheduler';
 
-// Ensure the JobManager exists but does not auto-start anything.
-beforeAll(() => {
-  // Nothing to start by default. Suites that need jobs will start them explicitly.
-});
+// Ensure jobs never auto-start during unit tests
+process.env.DISABLE_JOBS = '1';
+process.env.NODE_ENV = 'test';
 
-// Keep tests isolated: prevents job re-registration & stale singletons
+// Run before each test file
 beforeEach(() => {
-  vi.resetModules();
-  if (!process.env.LOG_LEVEL) process.env.LOG_LEVEL = 'fatal';
-});
-
-// Clean up after *every* test to prevent cross-test contamination.
-afterEach(async () => {
-  try {
-    await stopAllJobs();
-  } catch {}
-  resetJobsForTests?.();
-  resetBusForTests();
-  const jm = getJobManagerForTests?.();
-  jm?.resetForTests?.();
+  JobManager.instance().stopAll();
+  JobManager.instance().resetForTests?.();
+  resetSchedulerForTests?.();
+  resetBusForTests?.();
 });
