+++ apps/api/tsconfig.json
@@ -6,7 +6,7 @@
     "rootDir": ".",
     "declaration": true,
     "declarationMap": true,
-    "types": ["node", "vitest"]
+    "types": ["node", "vitest/globals"]
   },
   "include": ["src", "vitest.config.ts"]
 }
