+++ packages/audit/src/index.ts
@@ -1,1 +1,1 @@
-export * from './reader.js';
+export * from './audit.js';
