+++ packages/audit/tsconfig.json
@@ -4,7 +4,8 @@
     "outDir": "dist",
     "rootDir": ".",
     "declaration": true,
-    "declarationMap": true
+    "declarationMap": true,
+    "emitDeclarationOnly": false
   },
   "include": ["src", "vitest.config.ts"]
 }
